+++ scripts/scil_compute_fodf_metrics.py
@@ -37,7 +37,7 @@ from dipy.direction.peaks import reshape_peaks_for_visualization
 from scilpy.io.utils import (add_overwrite_arg, add_sh_basis_args,
                              add_processes_arg,
                              assert_inputs_exist, assert_outputs_exist)
-from scilpy.reconst.multi_process import peaks_from_sh, maps_from_sh
+from scilpy.reconst.multi_processes import peaks_from_sh, maps_from_sh
 
 
 def _build_arg_parser():
